+++ crates/datastore/src/locking_tx_datastore/mod.rs
@@ -8,6 +8,8 @@ mod sequence;
 pub mod state_view;
 pub use state_view::{IterByColEqTx, IterByColRangeTx};
 pub mod delete_table;
+mod replay;
+pub use replay::ErrorBehavior;
 mod tx;
 pub use tx::{NumDistinctValues, TxId};
 mod tx_state;
